Simplification of validator based on features from Test Bed release 1.11.0.
