+++ pom.xml
@@ -13,7 +13,7 @@
     <version>1.0-SNAPSHOT</version>
     <properties>
         <!-- Dependency and plugin versions. -->
-        <version.com.gitb>1.10.0</version.com.gitb>
+        <version.com.gitb>1.11.0</version.com.gitb>
         <version.org.apache.cxf>3.4.0</version.org.apache.cxf>
         <version.com.spotify.dockerfile-maven-plugin>1.4.12</version.com.spotify.dockerfile-maven-plugin>
         <version.maven-compiler-plugin>3.8.1</version.maven-compiler-plugin>
