+++ src/main/java/eu/europa/ec/itb/validator/eark/gitb/ServiceConfig.java
@@ -23,9 +23,6 @@ public class ServiceConfig {
     @Autowired
     ValidationServiceImpl validationServiceImplementation;
 
-    @Autowired
-    ProcessingServiceImpl processingServiceImplementation;
-
     /**
      * JSON serialiser/deserialiser.
      *
@@ -47,20 +44,6 @@ public class ServiceConfig {
         endpoint.setServiceName(new QName("http://www.gitb.com/vs/v1/", "ValidationService"));
         endpoint.setEndpointName(new QName("http://www.gitb.com/vs/v1/", "ValidationServicePort"));
         endpoint.publish("/validation");
-        return endpoint;
-    }
-
-    /**
-     * The CXF endpoint that will serve processing service calls.
-     *
-     * @return The endpoint.
-     */
-    @Bean
-    public Endpoint processingService() {
-        EndpointImpl endpoint = new EndpointImpl(cxfBus, processingServiceImplementation);
-        endpoint.setServiceName(new QName("http://www.gitb.com/ps/v1/", "ProcessingServiceService"));
-        endpoint.setEndpointName(new QName("http://www.gitb.com/ps/v1/", "ProcessingServicePort"));
-        endpoint.publish("/processing");
         return endpoint;
     }
